scripts/requirements.txt
numpy
pillow
facenet-pytorch
imgaug
torch
torchvision
opencv-python
joblib
mysql-connector-python
scikit-learn
python-dotenv
